% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_dat.R
\docType{data}
\name{alpha_dat}
\alias{alpha_dat}
\title{Elongation factor eEF1B subnit alpha}
\description{
Originally published in:
Bondarchuk, T. V., Shalak, V. F., Lozhko, D. M., Fatalska, A., Szczepanowski, R. H., 
Liudkovska, V., Tsuvariev, O. Y., Dadlez, M., El’skaya, A. V., & Negrutskii, B. S. (2022). 
Quaternary organization of the human eEF1B complex reveals unique multi-GEF domain assembly. 
Nucleic Acids Research, 50(16), 9490–9504. https://doi.org/10.1093/nar/gkac685
}
\references{
\url{https://doi.org/10.1093/nar/gkac685}
}
\author{
Bondarchuk, T. V., Shalak, V. F., Lozhko, D. M., Fatalska, A., Szczepanowski, R. H., 
Liudkovska, V., Tsuvariev, O. Y., Dadlez, M., El’skaya, A. V., & Negrutskii, B. S.
}
\keyword{data}
