% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_aggregated_difference_dataset.R
\name{create_aggregated_diff_uptake_dataset}
\alias{create_aggregated_diff_uptake_dataset}
\title{Calculates aggregated uptake difference for peptide pool}
\usage{
create_aggregated_diff_uptake_dataset(diff_uptake_dat)
}
\arguments{
\item{diff_uptake_dat}{differential uptake data, 
product of e.q. \code{\link{create_diff_uptake_dataset}}}
}
\value{
a \code{\link{data.frame}} object
}
\description{
This is a wrapper function for \code{\link{calculate_aggregated_diff_uptake}},
which calculates aggregated differential uptake for only
one time point. This function returns values for all 
time points from `diff_uptake_dat`.
}
\examples{
diff_uptake_dat <- create_diff_uptake_dataset(alpha_dat)
head(create_aggregated_diff_uptake_dataset(diff_uptake_dat))

}
