% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_residue_positions.R
\name{get_residue_positions}
\alias{get_residue_positions}
\title{Get residue positions}
\usage{
get_residue_positions(dat)
}
\arguments{
\item{dat}{data imported by the \code{\link{read_hdx}} function.}
}
\value{
a \code{data.frame} object.
}
\description{
Get residue positions
}
\details{
Prepares data table for high-resoultion values. It creates a long data frame 
with each position in the sequence accompanied by the amino acid placed in this
position.
}
\examples{
get_residue_positions(alpha_dat)

}
